move dashboard component into dashboard folder
