+++ client2/src/app/dashboard/dashboard-router/dashboard-router.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
 import {RouterModule, Routes} from "@angular/router";
-import {DashboardComponent} from "../dashboard.component";
+import {DashboardComponent} from "../dashboard/dashboard.component";
 import {AuthGuard} from "../../auth/auth-guard.service";
 import {DeviceOverviewComponent} from "../device-overview/device-overview.component";
