+++ client2/src/app/dashboard/dashboard.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
-import {DashboardComponent} from "./dashboard.component";
+import {DashboardComponent} from "./dashboard/dashboard.component";
 import {DashboardRouterModule} from "./dashboard-router/dashboard-router.module";
 import {AuthGuard} from "../auth/auth-guard.service";
 import {DeviceOverviewComponent} from "./device-overview/device-overview.component";
